m68k: Add support for uC5282 board
This patch adds support for emulation of the uC5282 board as
a new board in the m68k category. The machine can be invoked
using the '-M uc5282' option for qemu-system-m68k. It makes
use of the m5208 CPU for this emulation. This machine will be
built automatically when configuring and building qemu for
the m68k architecture.

Signed-off-by: Harrison Edward Gerber <hgerber@uccs.edu>